pyproject.toml
[tool.poetry]
name = "logging-end-to-end-test-tool"
version = "0.1.0"
description = ""
authors = ["Ryan Faircloth <ryan@dss-i.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^7.1.2"
PyYAML = "^6.0"
Jinja2 = "^3.1.2"
Faker = "^13.15.1"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
